mirror trxcllnt/js-data-refactor#4ba11bfb5ca54f34ac1a45703e035e66a9aa73e2

+++ fb/Schema.d.ts
@@ -436,7 +436,7 @@ export declare namespace org.apache.arrow.flatbuf {
          * @param {Array.<number>} data
          * @returns {flatbuffers.Offset}
          */
-        static createTypeIdsVector(builder: flatbuffers.Builder, data: number[] | Uint8Array): flatbuffers.Offset;
+        static createTypeIdsVector(builder: flatbuffers.Builder, data: number[] | Int32Array): flatbuffers.Offset;
         /**
          * @param {flatbuffers.Builder} builder
          * @param {number} numElems
